refactor: Move compact_in_parallel and compact_one_partition to lib and make more general
Cold compaction is going to use these too.
